Refactor Workflow dependencies + add parse_with_id for list/dict parsing
